BLD: fix ethereum UT + minor reorg
