+++ test/ethereum/utils.js
@@ -1,5 +1,8 @@
 const web3Utils = require('web3-utils');
 const crypto = require('../../src/common/cryptography');
+const DB_PROVIDER = require('../../src/core/core_server_mock/data/provider_db');
+const DbUtils = require('../../src/common/DbUtils');
+
 
 function runSelectionAlgo(secretContractAddress, seed, nonce, balancesSum, balances, workers) {
   const hash = web3Utils.soliditySha3(
@@ -105,3 +108,49 @@ module.exports.createDataForSelectionAlgorithm = function() {
   };
 };
 
+module.exports.transformStatesListToMap = (statesList) =>  {
+  const statesMap = {};
+  for (let i = 0; i < statesList.length; ++i) {
+    const address = statesList[i].address;
+    if (!(address in statesMap)) {
+      statesMap[address] = {};
+    }
+    const key = statesList[i].key;
+    const delta = statesList[i].data;
+    statesMap[address][key] = delta;
+  }
+  return statesMap;
+};
+
+module.exports.PROVIDERS_DB_MAP = this.transformStatesListToMap(DB_PROVIDER);
+
+// add the whole DB_PROVIDER as a state in ethereum. ethereum must be running for this worker
+module.exports.setEthereumState = async (api, web3, workerAddress, workerEnclaveSigningAddress) => {
+  for (const address in this.PROVIDERS_DB_MAP) {
+    const secretContractData = this.PROVIDERS_DB_MAP[address];
+    const addressInByteArray = address.split(',').map(function(item) {
+      return parseInt(item, 10);
+    });
+
+    const hexString = '0x' + DbUtils.toHexString(addressInByteArray);
+    const codeHash = crypto.hash(secretContractData[-1]);
+    const firstDeltaHash = crypto.hash(secretContractData[0]);
+    const outputHash = web3.utils.randomHex(32);
+    const gasUsed = 5;
+    const optionalEthereumData = '0x00';
+    const optionalEthereumContractAddress = '0x0000000000000000000000000000000000000000';
+    await api.deploySecretContract(hexString, codeHash, codeHash, firstDeltaHash, optionalEthereumData,
+      optionalEthereumContractAddress, gasUsed, workerEnclaveSigningAddress, {from: workerAddress});
+
+    let i = 1;
+
+    while (i in secretContractData) {
+      const taskId = web3.utils.randomHex(32);
+      const delta = secretContractData[i];
+      const stateDeltaHash = crypto.hash(delta);
+      await api.commitReceipt(hexString, taskId, stateDeltaHash, outputHash, optionalEthereumData, optionalEthereumContractAddress, gasUsed,
+        workerEnclaveSigningAddress, {from: workerAddress});
+      i++;
+    }
+  }
+};
